Add new method to fetch schedule, keeping the old one
New method needed as the APIs now return an URL, which then returns the schedule in XML
format (which is why this commit also adds a dependency to parse the XML).
Since this mechanism is horrible, I'm keeping the old code, in case they change their minds.
Also, the XML is send over http, so I had to set the NSAllowsArbitraryLoads flag.